Use `h` templating function
- Issue: https://git.chen.so/tina-next-ts/i/2956ad04253d830768cf07430c2f6cb6f9ed0151
