+++ app/layout.ts
@@ -1,0 +1,51 @@
+import type React from "react";
+import { createElement as h } from "react";
+import Link from "./components/Link";
+import "./styles/globals.css";
+import global from "../content/global/index.json";
+import { Box } from "../styled-system/jsx";
+
+export default function RootLayout({
+	children,
+}: {
+	children: React.ReactNode;
+}) {
+	return h(
+		"html",
+		{ lang: "en" },
+		h(
+			Box,
+			{
+				as: "body",
+				margin: "3rem",
+				color: { base: "token(colors.neutral.950)", _osDark: "white" },
+				backgroundColor: {
+					base: "white",
+					_osDark: "token(colors.neutral.950)",
+				},
+			},
+			[
+				h(
+					Box,
+					{
+						as: "header",
+						css: {
+							"& > a": {
+								fontSize: "lg",
+								textDecoration: "none",
+								marginX: 4,
+								"&:first-of-type": { marginLeft: 0 },
+								"&:last-of-type": { marginRight: 0 },
+							},
+						},
+					},
+					global.header.nav.map((link, i) => [
+						h(Link, { key: link.href, href: link.href }, link.label),
+						i === global.header.nav.length - 1 ? "" : " | ",
+					]),
+				),
+				h("main", { children }),
+			],
+		),
+	);
+}
